% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kth_largest_element.R
\name{kth_largest_element}
\alias{kth_largest_element}
\title{k-th largest element of a vector}
\usage{
kth_largest_element(nums, k)
}
\arguments{
\item{nums, }{a numeric vector where we would like to find elements. This function is trying to find a target in a group of numbers so this parameter is named as nums.}

\item{k, }{a numeric value that specifies the k-th largest element in the vector. This function is trying to find the k-th largest element in a vector so this parameter is named as k.}
}
\value{
a numeric value which is the k-th largest element in the vector
}
\description{
This function returns the k-th largest element in a vector with two parameters, nums and k on a 1-index basis.
If k is smaller than 1, this function will show a warning message and set k to 1 automatically.
Also if k is larger than the length of the vector, this function will show a warning message,
and set k to the length of the vector automatically.
}
\examples{
kth_largest_element(c(5, 4, 8, 7, 3, 6, 9, 1, 2, 0), 1)
kth_largest_element(c(3, 7, 9, 3), 2)
}
